+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.4.0 (2026-06-29)
+
+
+### Features
+
+* Define correct default values for application config
+* update comments and tests for policies
+
 ## 0.3.0 (2026-06-25)